LaptopAdvisor/recommendation_engine/requirements.txt
flask==2.3.3
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
mysql-connector-python==8.0.33
python-dotenv==1.0.0
scipy==1.10.1
